Move encodeEcdsaSigToValue to signer.go
This function is also needed in working with Windows certificate store
ECDSA signatures.
